feat: adiciona funcionalidade de poder editar o arquivo .env

+++ infra/setup.sh
@@ -9,38 +9,43 @@ echo "Verificando Docker..."
 if ! command -v docker >/dev/null 2>&1; then
     echo "Instalando Docker..."
     sudo apt install docker.io -y
-    else
-        echo "Docker ja esta instalado."
+else
+    echo "Docker ja esta instalado."
 fi
 
 echo "Verificando Docker Compose..."
 if ! docker compose version >/dev/null 2>&1; then
     echo "Instalando Docker Compose..."
     sudo apt install docker-compose -y
-    else
-        echo "Docker Compose já esta instalado."
+else
+    echo "Docker Compose já esta instalado."
 fi
 
 DIR_BASE="/opt/futour-sight"
 DIR_REPO="${DIR_BASE}/repo"
 
+MANTERENV="N"
+
 if [[ -d "$DIR_BASE" ]]; then
-    sudo rm -rf "$DIR_BASE"
+    read -p "Deseja manter o .env? (S/N): " MANTERENV
+    if [[ "$MANTERENV" == "S" || "$MANTERENV" == "s" ]]; then
+        sudo rm -rf "$DIR_REPO"
+    else
+        sudo rm -rf "$DIR_BASE"
+    fi
 fi
 
 sudo mkdir -p "$DIR_BASE"
 
 echo "Copiando repositório..."
-sudo cp -r "$HOME/FuTour-Sight/." "$DIR_REPO"
+sudo rsync -avh --exclude='node_modules/' --info=progress2 "$HOME/FuTour-Sight/." "$DIR_REPO"
 
 echo "Copiando Docker Compose..."
 sudo cp "$DIR_REPO/infra/dockers/docker-compose.yml" "$DIR_BASE/docker-compose.yml"
 
-echo "Copiando .env"
-sudo cp "$DIR_REPO/infra/env/.env.exemplo" "$DIR_BASE/.env"
-
-echo "Criando configuração de charset do MySQL..."
-sudo tee "$DIR_REPO/infra/mysql.cnf" > /dev/null << 'MYSQLCNF'
+if [[ "$MANTERENV" != "S" && "$MANTERENV" != "s" ]]; then
+    echo "Criando configuração de charset do MySQL..."
+    sudo tee "$DIR_REPO/infra/mysql.cnf" > /dev/null << 'MYSQLCNF'
 [mysqld]
 character-set-server=utf8mb4
 collation-server=utf8mb4_unicode_ci
@@ -52,16 +57,19 @@ default-character-set=utf8mb4
 default-character-set=utf8mb4
 MYSQLCNF
 
-echo "Arquivo .env criado em: $DIR_BASE/.env"
+    echo "Copiando .env"
+    sudo cp "$DIR_REPO/infra/env/.env.exemplo" "$DIR_BASE/.env"
+    echo "Arquivo .env criado em: $DIR_BASE/.env"
 
-echo ""
-read -p "Gostaria de preencher agora? (S/N): " RESPOSTA
+    echo ""
+    read -p "Gostaria de preencher agora? (S/N): " RESPOSTA
 
-if [[ "$RESPOSTA" == "S" || "$RESPOSTA" == "s" ]]; then
-    "$DIR_REPO/infra/editarEnv.sh"
+    if [[ "$RESPOSTA" == "S" || "$RESPOSTA" == "s" ]]; then
+        "$DIR_REPO/infra/editarEnv.sh"
     else
-    echo "Para preencher o .env, execute o script:"
-    echo "$DIR_REPO/infra/editarEnv.sh"
+        echo "Para preencher o .env, execute o script:"
+        echo "$DIR_REPO/infra/editarEnv.sh"
+    fi
 fi
 
 echo "Setup concluido!"
